add cloudwatch log export
this exposes the cloud watch log export functionality or mysql rds
instances. We already support setting the required role (so cloudwatch
can access the logs on the rds instance) as well as specifying custom
option and parameter groups (required to enable some types of logs e.g.
audit or slowlog) on the instance itself.

the cloudwatch log groups need to be created prior to the rds instance
itself. The reason is that creating the rds instance with enabled
exports will implicitly create the log groups at the AWS API level. In
that case terraform is not aware about the already existing log groups
and will not be able to manage them (to set the kms key or the retention
days). The dependency ensures the existance of the log groups before
enabling the export feature.

this additionally adds creation of param and option groups for the
instance if necessary. these will only be created if the user did
not specify an option/param group to use via a parameter and applied
settings require their creation.